refactor (potential) generics into separate fpa package
