+++ src/it/scala/contact/ContactServerSpec.scala
@@ -1,22 +1,23 @@
 package contact
 
 import cats.effect.IO
+
 import io.circe._
 import io.circe.literal._
 import io.circe.optics.JsonPath._
+
 import org.http4s.circe._
 import org.http4s.client.blaze._
 import org.http4s.server.blaze._
 import org.http4s.server.{Server => Http4sServer}
 import org.http4s._
+
 import org.scalatest._
 
-import config._
-import db._
-import repository._
-import service._
+import fpa._
 
 class ContactServerSpec extends WordSpec with Matchers with BeforeAndAfterAll {
+
   private lazy val client = Http1Client[IO]().unsafeRunSync()
 
   private lazy val config = Config.load("test.conf").unsafeRunSync()
