+++ src/main/scala/fpa/Util.scala
@@ -1,12 +1,10 @@
-package contact
+package fpa
 
-package object util {
+object stream {
 
-  object stream {
     import fs2.Stream
+
     implicit class EffectOps[F[_], A](fa: F[A]) {
       def stream: Stream[F, A] = Stream.eval(fa)
     }
-  }
-
 }
